Fix the "Art Board Too Small" Bug
Inches were not being converted to millimeters when run in Fusion 360 but were being converted in the test framework. It turns out the object returned by the getWorkpiece() is not a copy of the workpiece and is not writeable. Assignments are silently ignored by the execution environment in Fusion 360.

Fix is to manually make a copy of the workpiece.