Fix for rapid moves no longer provided by Fusion 360
Rapid moves were removed from the free version fo Fusion 360. This fix changes the way that path detection works by only drawing paths for moves when the laser is on. The laser state is detected with the `onPower` event instead of deducing it from the rapid moves between shapes.

Also updated the unit test framework to emit onPower events. It doesnt do it exactly how Fusion360 does it but the processor code is robust against the difference.